chore: rename sidebar-button directory
